Add the directory 0x0B-malloc_freeand updated READMEs

+++ 0x0B-malloc_free/0-create_array.c
@@ -21,7 +21,7 @@ char *create_array(unsigned int size, char c)
 	else
 	{
 		pt_ar = malloc(sizeof(int) * size);
-		value = (int)(sizeof(int) * size);
+		value = sizeof(int) * size;
 		if (pt_ar == NULL)
 		{
 			return (NULL);
